Update gradle Add minting functionality

+++ app/src/main/java/at/mintech/nftmaker/domain/entities/MintParams.kt
@@ -1,7 +1,9 @@
 package at.mintech.nftmaker.domain.entities
 
+import java.math.BigInteger
+
 data class MintParams(
     val address: String,
-    val tokenId: Int,
+    val tokenId: BigInteger,
     val nftUrl: String
 )
